Replaced pages with activeTab query parameter
