--- apps/example-orchestrator-ui/pages/subscriptions/active.tsx
@@ -1,10 +1,0 @@
-import {
-    SubscriptionsOverviewPage,
-    SubscriptionsTabType,
-} from '../../components/SubscriptionsPage/SubscriptionsPage';
-
-export default function ActiveSubscriptionsPage() {
-    return (
-        <SubscriptionsOverviewPage activeTab={SubscriptionsTabType.ACTIVE} />
-    );
-}
