--- apps/example-orchestrator-ui/pages/subscriptions/non-active.tsx
@@ -1,12 +1,0 @@
-import {
-    SubscriptionsOverviewPage,
-    SubscriptionsTabType,
-} from '../../components/SubscriptionsPage/SubscriptionsPage';
-
-export default function ActiveSubscriptionsPage() {
-    return (
-        <SubscriptionsOverviewPage
-            activeTab={SubscriptionsTabType.NON_ACTIVE}
-        />
-    );
-}
